+++ Banking/DBManager/Impl/AccountManagerImpl.cs
@@ -7,12 +7,12 @@ namespace Banking.DBManager.Impl {
 
     public interface IAccountManagerImpl {
         public void AddAccount(int custId, int accNo, char type,
-            double balance);
+            decimal balance);
         public Account GetAccountByAccountNumber(int accNo);
         public List<int> GetAccountNumbersForCustomer(string custId);
-        public void Deposit(int accNo, double amount, string comment);
-        public void WithDraw(int accNo, double amount, string comment);
-        public void Transfer(int srcNo, int destNo, double amount,
+        public void Deposit(int accNo, decimal amount, string comment);
+        public void WithDraw(int accNo, decimal amount, string comment);
+        public void Transfer(int srcNo, int destNo, decimal amount,
             string comment);
     }
 
@@ -33,11 +33,11 @@ namespace Banking.DBManager.Impl {
                 AccountNumber = (int)reader["AccountNumber"],
                 AccountType = reader["AccountType"].ToString().ToCharArray()[0],
                 CustomerID = (int)reader["CustomerID"],
-                Balance = (double)reader["Balance"]
+                Balance = (decimal)reader["Balance"]
             };
         }
 
-        public void AddAccount(int custId, int accNo, char type, double balance) {
+        public void AddAccount(int custId, int accNo, char type, decimal balance) {
             using (var conn = GetConnection())
             {
                 conn.Open();
@@ -96,7 +96,7 @@ from {TableName} where CustomerID = @CustId";
             return accNos;
         }
 
-        private void Deposit(int accNo, double amount, SqlCommand command)
+        private void Deposit(int accNo, decimal amount, SqlCommand command)
         {
             command.CommandText = $@"update {TableName}
 set Balance += @Amount
@@ -108,7 +108,7 @@ where AccountNumber = @AccNo";
             command.ExecuteNonQuery();
         }
 
-        public void Deposit(int accNo, double amount, string comment) {
+        public void Deposit(int accNo, decimal amount, string comment) {
             using (var conn = GetConnection())
             {
                 conn.Open();
@@ -131,7 +131,7 @@ where AccountNumber = @AccNo";
             }
         }
 
-        private void WithDraw(int accNo, double amount, SqlCommand command)
+        private void WithDraw(int accNo, decimal amount, SqlCommand command)
         {
             command.CommandText = $@"update {TableName}
 set Balance -= @Amount
@@ -143,7 +143,7 @@ where AccountNumber = @AccNo";
             command.ExecuteNonQuery();
         }
 
-        public void WithDraw(int accNo, double amount, string comment) {
+        public void WithDraw(int accNo, decimal amount, string comment) {
             using (var conn = GetConnection())
             {
                 conn.Open();
@@ -168,7 +168,7 @@ where AccountNumber = @AccNo";
             }
         }
 
-        private void Transfer(int srcNo, int destNo, double amount,
+        private void Transfer(int srcNo, int destNo, decimal amount,
             SqlCommand command)
         {
             command.CommandText = $@"update {TableName}
@@ -184,7 +184,7 @@ set Balance -= @Amount where AccountNumber = @DestNo";
             command.ExecuteNonQuery();
         }
 
-        public void Transfer(int srcNo, int destNo, double amount,
+        public void Transfer(int srcNo, int destNo, decimal amount,
             string comment)
         {
             using (var conn = GetConnection())
